+++ helpers/Math.go
@@ -1,5 +1,5 @@
 package helpers
 
-func Between(v int, min int, max int) bool {
+func Between(v int8, min int8, max int8) bool {
 	return v >= min && v <= max
 }
